requirements.txt
asgiref
astroid
autopep8
boto3
botocore
certifi
cffi
charset-normalizer
colorama
cryptography
Django==3.0
django-cors-headers
django-cprofile-middleware
django-storages
djangorestframework==3.11.1
docutils
et-xmlfile
Faker
flake8
idna
importlib-metadata
isort
jdcal
jmespath
lazy-object-proxy
mccabe
mysqlclient
numpy
numpy-financial
openpyxl
pandas
pycodestyle
pycparser
pylint
python-dateutil
python-dotenv
pytz
s3transfer
selenium
six
sqlparse
text-unidecode
typed-ast
typing-extensions
urllib3
watchtower
wrapt
xlrd
zipp
